+++ Java/src/main/java/com/evolvetechnology/datematchers/FreeDateMatcherAggregator.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology;
+package com.evolvetechnology.datematchers;
 
 import java.time.LocalDateTime;
 import java.util.*;
