+++ Java/src/main/java/com/evolvetechnology/datematchers/FreeHolidayMatcher.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology;
+package com.evolvetechnology.datematchers;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
