+++ Java/src/main/java/com/evolvetechnology/datematchers/FreeWeekDayMatcher.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology;
+package com.evolvetechnology.datematchers;
 
 import java.time.DayOfWeek;
 import java.time.LocalDateTime;
