+++ Java/src/main/java/com/evolvetechnology/timecost_calculation/TimeCostCalculator.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology;
+package com.evolvetechnology.timecost_calculation;
 
 import java.time.LocalTime;
