+++ Java/src/main/java/com/evolvetechnology/vehicles/Car.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology.vehicle;
+package com.evolvetechnology.vehicles;
 
 public class Car implements Vehicle {
