+++ Java/src/main/java/com/evolvetechnology/vehicles/Foreign.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology.vehicle;
+package com.evolvetechnology.vehicles;
 
 public class Foreign implements Vehicle {
   @Override
