+++ Java/src/main/java/com/evolvetechnology/vehicles/Motorbike.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology.vehicle;
+package com.evolvetechnology.vehicles;
 
 public class Motorbike implements Vehicle {
